Make related changes to support async mesh.
Make related changes in zooshi to support making async LoadMesh.
Since now the LoadWorldDef may load mesh asynchronously, we need to make
sure the game menu is showing after all assets loaded.

Issue: b/30067916

Tested on mac.

+++ src/components/river.cpp
@@ -423,7 +423,7 @@ void RiverComponent::CreateRiverMesh(corgi::EntityRef& entity) {
   // Create the actual mesh objects, and stuff all the data we just
   // generated into it.
   Mesh* river_mesh =
-      new Mesh(river_verts.data(), static_cast<int>(river_verts.size()),
+      new Mesh(river_verts.data(), river_verts.size(),
                static_cast<int>(sizeof(NormalMappedVertex)), kMeshFormat);
 
   river_mesh->AddIndices(river_indices.data(),
